restructure client sources
