+++ internal/cli/views/basics/item/item.go
