+++ internal/cli/views/items/pair/show/page.go
@@ -1,7 +1,7 @@
 package show
 
 import (
-	"iwakho/gopherkeep/internal/cli/views/item"
+	"iwakho/gopherkeep/internal/cli/views/basics/item"
 
 	"github.com/charmbracelet/bubbles/list"
